+++ src/context.ts
@@ -1,0 +1,8 @@
+import * as React from 'react';
+
+const SliderContext = React.createContext({
+  min: 0,
+  max: 0,
+});
+
+export default SliderContext;
